+++ devtools/android/src/test/kotlin/com/maximbircu/devtools/android/readers/soruces/yaml/YamlDevToolsReaderTest.kt
@@ -2,6 +2,8 @@ package com.maximbircu.devtools.android.readers.soruces.yaml
 
 import com.maximbircu.devtools.android.BaseTest
 import com.maximbircu.devtools.common.core.DevTool
+import com.maximbircu.devtools.common.presentation.tools.text.TextTool
+import com.maximbircu.devtools.common.presentation.tools.time.TimeTool
 import com.maximbircu.devtools.common.presentation.tools.toggle.ToggleTool
 import io.mockk.every
 import io.mockk.mockk
@@ -25,12 +27,17 @@ class YamlDevToolsReaderTest : BaseTest() {
 
     @Test
     fun `returns proper dev tool`() {
-        enqueueTools(mapOf("toggle-tool" to ToggleTool()))
+        val expectedTools: Map<String, DevTool<*>> = mapOf(
+            "toggle-tool" to ToggleTool(),
+            "text-tool" to TextTool(),
+            "time-tool" to TimeTool()
+        )
+        enqueueTools(expectedTools)
         val reader = YamlDevToolsReader(mockk(relaxed = true), "".byteInputStream())
 
-        val tools = reader.getDevTools()
+        val actualTools = reader.getDevTools()
 
-        assertEquals(tools.getValue("toggle-tool").key, "toggle-tool")
+        assertEquals(expectedTools, actualTools)
     }
 
     @Suppress("UNCHECKED_CAST")