+++ devtools/common/src/commonTest/kotlin/com/maximbircu/devtools/common/DevToolsParserImplTest.kt
@@ -1,36 +1,49 @@
 package com.maximbircu.devtools.common
 
 import com.maximbircu.devtools.common.core.DevTool
-import com.maximbircu.devtools.common.core.createTool
 import com.maximbircu.devtools.common.core.reader.DevToolsReader
 import com.maximbircu.devtools.common.core.reader.DevToolsSource
 import com.maximbircu.devtools.common.mvp.BaseTest
+import com.maximbircu.devtools.common.presentation.tools.group.GroupTool
+import com.maximbircu.devtools.common.presentation.tools.text.TextTool
+import com.maximbircu.devtools.common.presentation.tools.time.TimeTool
+import com.maximbircu.devtools.common.presentation.tools.toggle.ToggleTool
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
 class DevToolsParserImplTest : BaseTest() {
     @Test
     fun `returns a proper map of tools from the provided sources`() {
-        val source1Tools = mapOf<String, DevTool<Any>>(
-            "first-source-first-tool" to createTool(),
-            "first-source-second-tool" to createTool()
+        val source1Tools = mapOf<String, DevTool<*>>(
+            "first-source-first-tool" to TimeTool(),
+            "first-source-second-tool" to TextTool()
         )
-
-        val source2Tools = mapOf<String, DevTool<Any>>(
-            "second-source-first-tool" to createTool(),
-            "second-source-second-tool" to createTool()
+        val source2Tools = mapOf<String, DevTool<*>>(
+            "second-source-first-tool" to TextTool(),
+            "second-source-second-tool" to ToggleTool(),
+            "group-tool" to GroupTool(tools = mapOf("child-tool" to TextTool()))
         )
+        val sources = listOf(createSource(source1Tools), createSource(source2Tools))
+        val parser = DevToolsParser.create(sources)
 
-        val sourcesList = listOf(createSource(source1Tools), createSource(source2Tools))
+        val tools = parser.getDevTools()
 
-        val parser = DevToolsParser.create(sourcesList)
+        assertEquals(source1Tools + source2Tools, tools)
 
-        assertEquals(parser.getDevTools(), source1Tools + source2Tools)
+        assertEquals("first-source-first-tool", tools.getValue("first-source-first-tool").key)
+        assertEquals("first-source-second-tool", tools.getValue("first-source-second-tool").key)
+        assertEquals("second-source-first-tool", tools.getValue("second-source-first-tool").key)
+        assertEquals("second-source-second-tool", tools.getValue("second-source-second-tool").key)
+        assertEquals("second-source-second-tool", tools.getValue("second-source-second-tool").key)
+
+        val groupTool: GroupTool = tools.getValue("group-tool") as GroupTool
+        assertEquals("group-tool", tools.getValue("group-tool").key)
+        assertEquals("group-tool.child-tool", groupTool.tools.getValue("child-tool").key)
     }
 
-    private fun createSource(tools: Map<String, DevTool<Any>>) = object : DevToolsSource {
+    private fun createSource(tools: Map<String, DevTool<*>>) = object : DevToolsSource {
         override fun getReader() = object : DevToolsReader {
-            override fun getDevTools(): Map<String, DevTool<Any>> = tools
+            override fun getDevTools(): Map<String, DevTool<*>> = tools
         }
     }
 }